Power rates are now doubles to support accuracy in the sub-mW range.

+++ trunk/org.qualitune.jouleunit.test/src/org/qualitune/jouleunit/test/PowerRateMock.java
@@ -55,7 +55,7 @@ public class PowerRateMock extends AbstractPowerRate {
 	 * @see
 	 * org.qualitune.eunit.profiler.EMeasuringValue#getChargingRateInMilliWatt()
 	 */
-	public long getPowerRate() {
+	public double getPowerRate() {
 		return powerRate;
 	}
