+++ trunk/profilers/org.qualitune.jouleunit.android/src/org/qualitune/jouleunit/android/AndroidPowerRate.java
@@ -62,7 +62,7 @@ public class AndroidPowerRate extends AbstractPowerRate {
 	 * @see
 	 * org.qualitune.eunit.profiler.EMeasuredValue#getChargingRateInMilliWatt()
 	 */
-	public long getPowerRate() {
+	public double getPowerRate() {
 		return (voltage * current) / 1000l;
 	}
