+++ trunk/profilers/org.qualitune.jouleunit.linux/src/org/qualitune/jouleunit/linux/LinuxPowerRate.java
@@ -201,7 +201,7 @@ public class LinuxPowerRate extends AbstractPowerRate {
 	 * @see
 	 * org.qualitune.eunit.profiler.EMeasuringValue#getChargingRateInMilliWatt()
 	 */
-	public long getPowerRate() {
+	public double getPowerRate() {
 		return getChargingRate() * getPresentVoltage() / 1000 * -1;
 	}
